Revert "fix: netty v3 record & mock"
This reverts commit 28ba00d94e5d16760154fb7f2de61487e00e1dde.
